Add files via upload
This is my firs project in HTML and JS.
I did it for programming course. The main objetive was create a management system for a shipping company.
You can enter either as a company or as an importer.
It has many different funcionalities (depending on the type of user)  such as create new trips, check pending requests, assign requests, check statistics, cancel pending requests, among others.
This code could and will be better, but I'm very proud of it!